requirements.txt
fastapi==0.111.1
passlib==1.7.4
pydantic==2.8.2
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
SQLAlchemy==2.0.31
starlette==0.38.2
uvicorn==0.30.3
